Added 2 new projects 	Email 	Admin
Junit complete to as close to 100% coverage as possible for suitable
items

sorted and cleanup of email server

every class in admin is new. They borrow highly from existing code to
display content

sql for some of the new classes proving difficult

the admin login system works with all users having a admin level. This
is used to limit tasks/features to certain clearance levels